backend/requirements.txt
# File: backend/requirements.txt
# Purpose: Production dependencies for the MacOS Agent backend

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP Client
httpx==0.27.2

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
asyncpg==0.30.0

# Cache
redis[hiredis]==5.2.0

# Logging
structlog==24.4.0
python-json-logger==3.2.1

# Async Tasks
celery==5.4.0

# Retry Logic
tenacity==9.0.0

# Testing
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code Quality
black==24.10.0
isort==5.13.2
mypy==1.13.0
pylint==3.3.2

# Monitoring (optional)
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
